внес правки после ревью
