+++ ImageFeedIos/Profile/ProfileService.swift
@@ -4,25 +4,6 @@ enum ProfileServiceError: Error {
     case invalidRequest
 }
 
-struct ProfileResult: Codable {
-    let username: String
-    let first_name: String
-    let last_name: String?
-    let bio: String?
-}
-
-struct Profile {
-    let username: String
-    let name: String
-    let bio: String
-    
-    init(profileResult: ProfileResult) {
-        self.username = "@" + profileResult.username
-        self.name = profileResult.first_name + " " + (profileResult.last_name ?? "")
-        self.bio = profileResult.bio ?? ""
-    }
-}
-
 final class ProfileService {
     static let shared = ProfileService()
     private init() {}
@@ -44,7 +25,7 @@ final class ProfileService {
     }
     
     private func makeProfileURL () -> URLRequest? {
-        guard let url = URL(string: "https://api.unsplash.com/me") else {
+        guard let url = URL(string: Constants.profileURL) else {
             assertionFailure("[ProfileService] Failed to create ProfileURL")
             return nil
         }
